% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_MeC}
\alias{rw_MeC}
\title{A markov sampler using Rcpp}
\usage{
rw_MeC(sigma, x0, N)
}
\arguments{
\item{sigma}{the variance}

\item{x0}{the Initial value}

\item{N}{Number of random numbers}
}
\value{
Returns random sequence and rejection probability
}
\description{
a random walk Metropolis sampler for generating the standard Laplace distribution
}
\examples{
\dontrun{
rw_MeC(5,7,500)
}
}
