% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multi.R
\name{select}
\alias{select}
\title{Multi-attribute Decision Based on Information Entropy using R}
\usage{
select(x)
}
\arguments{
\item{x}{the data matrix which the columns represent the plans and the rows represent the variables considered}
}
\value{
a vector of attribute values of plans which were sorted from small to large and choose the best plan
}
\description{
Multi-attribute Decision Based on Information Entropy using R
}
\examples{
\dontrun{
x<-matrix(rnorm(16),4)
select(x)
}
}
