\name{StatComp20022-package}
\alias{StatComp20022-package}
\docType{package}
\title{
  Multi-attribute Decision Based on Information Entropy.
}
\description{
  R  versions of select and savage.select function and Rcpp version of rw are provided in this package.
}
\author{
Xin Zhang

Maintainer: Xin Zhang <xinzhang32@mail.ustc.edu.cn>
}
\keyword{ package }
